final works

+++ routes/api/members.js
@@ -9,30 +9,38 @@ router.use(express.json())
 const Job = require('../../models/Job')
 
 const Member = require('../../models/Member')
-
+const Admin = require ('../../models/Admin')
 const Workshop = require('../../models/Workshop') //yara
 const Assessment = require('../../models/Assessment') 
 
 const objectId = require('mongoose').objectid //yara
-
-
+const jwt = require('jsonwebtoken')
+const tokenKey = require('../../config/keys').secretOrKey
+const bcrypt = require('bcryptjs');
+const passport = require('passport');//for auth trial
 
 router.get('', async (req, res) => {
-
+  console.log("in get all")
   const members = await Member.find()
-
   res.json({ data: members })
 
 });
 
 router.get('/getexperts' , async (req , res) => {
-  const members = await Member.find({expert : true})
-res.json(members);
-})
 
-router.get('./routes/api/assessments/:name' , async (req , res) => {
-const assessments = await Assessment.find({memberName : name})
+  const members = await Member.find({expert : true})
+  res.json(members);
+
+});
+
+router.get('./routes/api/assessments/:name' , passport.authenticate('jwt', { session: false }) ,  async (req , res) => {
+  const id = req.params._id;
+  const id2=admin.id
+  const id3=req.user.id
+  if(id3==id2 || id3==id) {  
+const assessments = await Assessment.find({expertName : name})
 res.json(assessments);
+} 
 });
 
 function Age(birthday) { // birthday is a date
@@ -189,7 +197,7 @@ function commen(array1, array2) {
 router.post('/', async (req, res) => {
 
   try {
-
+   
     const isValidated = validator.createValidation(req.body)
 
     if (isValidated.error) return res.status(400).send({ error: isValidated.error.details[0].message })
@@ -203,7 +211,6 @@ router.post('/', async (req, res) => {
       if (found)
         return res.status(400).send({ error: "email is already in use" });
       else {
-
         const newMember = await Member.create(req.body)
 
         newMember.Age = Age(req.body.birthDate)
@@ -212,7 +219,6 @@ router.post('/', async (req, res) => {
       }
     }
   }
-
   catch (error) {
 
     // We will be handling the error later
@@ -221,16 +227,20 @@ router.post('/', async (req, res) => {
   }
 
 });
-
-router.get("/:_id", (req, res) => {
-  const id = req.params._id;
-
-  Member.findById(id)
-
+router.get("/:id", passport.authenticate('jwt', { session: false })  , async (req, res) => {
+  const id = req.params.id;
+  console.log("in getid")
+  console.log(req.params.id)
+  console.log(req.user._id)
+  const admin = await Admin.findById(req.user._id )
+  let id2 = '';
+  if (admin)
+  id2 = admin._id
+  const id3=req.user._id
+  if(id3 == id2 || id3 == id) {   
+    Member.findById(id)
     .exec()
-
     .then(r => {
-
       if (r) {
         const ret = {};
         ret.data = r;
@@ -238,22 +248,36 @@ router.get("/:_id", (req, res) => {
         res.status(200).json(ret);
         res.status(200).json(r);
       } else {
-
         res
-
           .status(404)
           .json({ message: "Member Not Found" });
+
       }
     })
     .catch(err => {
-      return "Member Not Found"
+      res.json("Member Not found")
     });
+  }
+  else {
+    res.json("Not authorized");
+  }
 });
-router.put('/:_id', async (req, res) => {
+
+router.put('/:_id',passport.authenticate('jwt', { session: false }), async (req, res) => {
+  const id = req.params._id;
+  console.log("in put")
+  const admin = await Admin.findById(req.user._id )
+  let id2 = '';
+  if (admin)
+  id2 = admin._id
+  const id3=req.user._id 
+  console.log("put request")
+  console.log(req.user._id) ;
+  console.log(id2);
+  console.log(req.params._id);
+  if(id3 == id2 || id3 == id) {
   try {
-
-    const isValidated = validator.updateValidation(req.body)
-
+       const isValidated = validator.updateValidation(req.body)
     if (isValidated.error) return res.status(400).send({ error: isValidated.error.details[0].message })
 
     const username = req.body.userName;
@@ -265,45 +289,58 @@ router.put('/:_id', async (req, res) => {
       const found = await Member.findOne({ "email": email });
       if (found)
         return res.status(400).send({ error: "email is already in use" });
-
       else {
         const updated = await Member.findByIdAndUpdate(req.params._id, req.body)
-
         return res.send(updated)
-
-      }
+      }}
     }
-  }
   catch (err) {
     res.send(err)
   }
-
-
+}
+else {
+  res.json("Not authenticated")
+}
 });
 
 
 
-router.delete('/:_id', async (req, res) => {
-
-  try {
-
+router.delete('/:_id' , passport.authenticate('jwt', { session: false }), async (req, res) => {
+  const id = req.params._id;
+  console.log("in Delete")
+  const admin = await Admin.findById(req.user._id )
+  let id2 = '';
+  if (admin)
+  id2 = admin._id
+  const id3=req.user._id 
+  console.log(req.user._id) ;
+  console.log(id2);
+  console.log(req.params._id);
+  if(id3 == id2 || id3 == id) {
+  try { 
      const deletedMember = await Member.findOneAndDelete ({ _id : req.params._id})
      res.json({msg :'Member deleted successfully' , data : deletedMember})
 
-
-  }
-
+    }
   catch (error) {
 
     console.log(error)
-
-  }
-})
-
+  }}
+});
 
 //MAYAR
 router.put('/:id/addcourse', async(req,res) => {
-
+  const id = req.params._id;
+  console.log("in Delete")
+  const admin = await Admin.findById(req.user._id )
+  let id2 = '';
+  if (admin)
+  id2 = admin._id
+  const id3=req.user._id 
+  console.log(req.user._id) ;
+  console.log(id2);
+  console.log(req.params._id);
+  if(id3 == id2 || id3 == id) {
   const isValidated = validator.applyValidation(req.body)
   if (isValidated.error) return res.status(400).send({ error: isValidated.error.details[0].message });
 
@@ -316,10 +353,22 @@ router.put('/:id/addcourse', async(req,res) => {
   .then(doc => { return res.redirect(303, `/api/members/${req.params.id}`) })
   .catch(err => { console.log(err); return res.send(`Sorry, couldn't update a member with that id !`) });
   console.log('one')
-
+  }
+  else {
+    res.json("Not Authenticated")
+  }
 });
 router.put('/:id/addskills', async(req,res) => {
-
+  const id = req.params._id;
+  const admin = await Admin.findById(req.user._id )
+  let id2 = '';
+  if (admin)
+  id2 = admin._id
+  const id3=req.user._id 
+  console.log(req.user._id) ;
+  console.log(id2);
+  console.log(req.params._id);
+  if(id3 == id2 || id3 == id) {
   const isValidated = validator.applyValidation(req.body)
   if (isValidated.error) return res.status(400).send({ error: isValidated.error.details[0].message });
 
@@ -331,9 +380,22 @@ router.put('/:id/addskills', async(req,res) => {
   .exec()
   .then(doc => { return res.redirect(303, `/api/members/${req.params.id}`) })
   .catch(err => { console.log(err); return res.send(`Sorry, couldn't update a member with that id !`) });
+  }
+  else {
+    res.json("Not Authenticated")
+  }
 });
 router.put('/:id/addinterests', async(req,res) => {
-
+  const id = req.params._id;
+  const admin = await Admin.findById(req.user._id )
+  let id2 = '';
+  if (admin)
+  id2 = admin._id
+  const id3=req.user._id 
+  console.log(req.user._id) ;
+  console.log(id2);
+  console.log(req.params._id);
+  if(id3 == id2 || id3 == id) {
   const isValidated = validator.applyValidation(req.body)
   if (isValidated.error) return res.status(400).send({ error: isValidated.error.details[0].message });
 
@@ -345,6 +407,10 @@ router.put('/:id/addinterests', async(req,res) => {
   .exec()
   .then(doc => { return res.redirect(303, `/api/members/${req.params.id}`) })
   .catch(err => { console.log(err); return res.send(`Sorry, couldn't update a member with that id !`) });
+  }
+  else{
+    res.json("Not authenticated");
+  }
 });
 
-module.exports = router
+module.exports = router;
